refactor: optimize backend architecture, caching, and history persistence.

+++ Backend/src/main/java/com/LuckyHub/Backend/event/PaymentRefundEvent.java
@@ -4,14 +4,16 @@ import com.LuckyHub.Backend.entity.User;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
 
+import java.math.BigDecimal;
+
 @Getter
 public class PaymentRefundEvent extends ApplicationEvent {
     private final User user;
     private final String planName;
     private final String orderId;
-    private final int amount;
+    private final BigDecimal amount;
 
-    public PaymentRefundEvent(User user, String planName, String orderId, int amount) {
+    public PaymentRefundEvent(User user, String planName, String orderId, BigDecimal amount) {
         super(user);
         this.user = user;
         this.planName = planName;
